+++ dlcdb/core/admin/base_admin.py
@@ -2,20 +2,18 @@
 #
 # SPDX-License-Identifier: EUPL-1.2
 
-import csv
-import datetime
 from django.db.models import Q
 from django.contrib import admin
-from django.http import HttpResponseRedirect, HttpResponse
+from django.http import HttpResponseRedirect
 from django.urls import reverse
-from django.utils import timezone, dateformat
+from django.utils import timezone
 from django.db.models import Count
 from django.utils.http import urlencode
 from django.utils.html import format_html
 from django.urls import path
 
 
-from ..models import Device, Room, DeviceType, Supplier, Manufacturer, LentRecord, Record
+from ..models import Device, Room, DeviceType, Supplier, Manufacturer
 from ..utils.helpers import get_denormalized_user, get_icon_for_class
 
 
@@ -254,71 +252,6 @@ class RedirectToDeviceMixin(object):
         device_obj = Device.objects.get(id=obj.device_id)
 
         return HttpResponseRedirect(reverse("admin:core_device_change", args=[device_obj.pk]))
-
-
-class ExportCsvMixin:
-    @admin.action(description="Export selected as CSV")
-    def export_as_csv(self, request, queryset):
-        meta = self.model._meta
-        _now = dateformat.format(timezone.now(), "Y-m-d_H-i-s")
-        filename = f"dlcdb_export_{_now}_{meta.app_label}-{meta.model_name}.csv"
-
-        fieldnames = [field.name for field in Device._meta.fields]  # .get_fields()
-        RECORD_FIELDNAMES = [field.name for field in Record._meta.fields]
-
-        if meta.model is Device:
-            # fieldnames = [field.name for field in meta.fields]
-            fieldnames.extend(["room", "created_at"])
-            export_queryset = queryset
-        elif meta.model is LentRecord:
-            device_pks = queryset.values_list("device", flat=True)
-            export_queryset = Device.objects.filter(pk__in=device_pks)
-            fieldnames.extend(["person", "lent_desired_end_date", "room", "created_at"])
-
-        response = HttpResponse(content_type="text/csv")
-        response["Content-Disposition"] = f"attachment; filename={filename}"
-        writer = csv.DictWriter(
-            response,
-            fieldnames=fieldnames,
-            dialect="excel-tab",
-            # delimiter=';',
-            # quotechar='"',
-            quoting=csv.QUOTE_ALL,
-            extrasaction="raise",
-        )
-
-        writer.writeheader()
-        for obj in export_queryset:
-            row_data = {}
-            for field in fieldnames:
-                fieldname = field
-                fielddata = None
-
-                if field == "active_record":
-                    fielddata = obj.active_record.record_type if obj.active_record else "no record set"
-
-                if field in RECORD_FIELDNAMES:
-                    # elif field == 'record_created_at':
-                    #     fielddata = obj.active_record.created_at if obj.active_record else "no record set"
-                    # elif field == 'room':
-                    #     fielddata = getattr(obj.active_record.room, "number", None) if obj.active_record else "no record set"
-                    # elif field == 'person':
-                    #     fielddata = obj.active_record.person if obj.active_record.person else "no person set"
-                    fielddata = getattr(obj.active_record, field)
-                else:
-                    fielddata = getattr(obj, field)
-
-                if isinstance(fielddata, datetime.datetime):
-                    fielddata = f"{fielddata:%Y-%m-%d %H:%M}"
-
-                row_data.update({fieldname: fielddata})
-
-            writer.writerow(row_data)
-
-        # Does not work as we do not trigger a HttpResponseRedirect which could display that message.
-        # self.message_user(request, f'Export file "{filename}" created.', messages.SUCCESS)
-
-        return response
 
 
 class DeviceCountMixin: