+++ dlcdb/core/admin/device_admin.py
@@ -19,13 +19,14 @@ from django.utils.translation import gettext_lazy as _
 from simple_history.admin import SimpleHistoryAdmin
 
 from dlcdb.tenants.admin import TenantScopedAdmin
+from dlcdb.dataexchange.admin import ExportCsvMixin
 
 from .. import lifecycle
 from ..models import Device, Record
 from ..utils.helpers import get_superuser_list
 from .filters.duplicates_filter import DuplicateFilter
 from .filters.recordtype_filter import HasRecordFilter
-from .base_admin import SoftDeleteModelAdmin, CustomBaseModelAdmin, ExportCsvMixin, get_has_note_badge
+from .base_admin import SoftDeleteModelAdmin, CustomBaseModelAdmin, get_has_note_badge
 
 
 # class NoteInline(admin.TabularInline):